Treat BOM exclusions as node excludes
When a BOM is imported, we create dependency constraint for each
dependency declared in a `<dependencyManagement>` section. These
entries may define a set of `<exclusions>` for the target module.
We were already parsing and caching these constraint exclusions,
but they were being ignored when building the graph, due to the
non-transitive nature of dependency constraints.

With this change, we consider any `exclude` defined for a constraint
as applying to the _node_ of the target module, and not tied to the
edge that brings in the constraint. This makes BOM import function
in a more Maven-like way, since these exclusions will not be ignored
when another dependency (with no define excludes) points at the
same module.

At this stage, it is not possible to define `excludes` for a
dependency defined in a build script, nor can they be persisted to
a .module metadata file. For now, constraint exclusions are exlusive
to BOM import.